+++ Locales/enUS.lua
@@ -55,6 +55,10 @@ L["Show minimap button"] = "Show minimap button"
 L["Allow ElvUI to skin this addon"] = "Allow ElvUI to skin this addon"
 L["Requires a /reload to take effect."] = "Requires a /reload to take effect."
 
+L["Author: %s"] = "Author: %s"
+L["More: %s"] = "More: %s"
+L["Version: %s"] = "Version: %s"
+
 L["Currencies"] = "Currencies"
 L["Mounts"] = "Mounts"
 L["Transmog"] = "Transmog"